requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.3
gspread==6.1.2
google-auth==2.29.0
google-cloud-texttospeech==2.16.3
google-cloud-speech==2.26.0
python-dotenv==1.0.1
